requirements.txt
streamlit==1.43.2
snowflake-connector-python==3.14.0
sseclient-py==1.8.0
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
